optimizing deferred renderer... lights not working any more xD --working on it
git-svn-id: https://happy-engine.googlecode.com/svn/trunk@78 5aa1a52c-6ec2-ce28-c747-3d7cbe9a755d

+++ src/HappyEngine/FPSGraph.cpp
@@ -235,7 +235,7 @@ uint FPSGraph::getAverageFPS() const
 
 	if (static_cast<uint>(m_GameTime) > 60)
 	{
-		avFPS /= (60 / m_Interval);
+		avFPS /= static_cast<uint>(60 / m_Interval);
 	}
 	else
 	{
